+++ aws-java-sdk-chimesdkmessaging/src/main/java/com/amazonaws/services/chimesdkmessaging/AmazonChimeSDKMessaging.java
@@ -30,8 +30,8 @@ import com.amazonaws.services.chimesdkmessaging.model.*;
  * The Amazon Chime SDK Messaging APIs in this section allow software developers to send and receive messages in custom
  * messaging applications. These APIs depend on the frameworks provided by the Amazon Chime SDK Identity APIs. For more
  * information about the messaging APIs, see <a
- * href="https://docs.aws.amazon.com/chime/latest/APIReference/API_Operations_Amazon_Chime_SDK_Messaging">Amazon Chime
- * SDK messaging</a>
+ * href="https://docs.aws.amazon.com/chime/latest/APIReference/API_Operations_Amazon_Chime_SDK_Messaging.html">Amazon
+ * Chime SDK messaging</a>.
  * </p>
  */
 @Generated("com.amazonaws:aws-java-sdk-code-generator")
@@ -1423,6 +1423,8 @@ public interface AmazonChimeSDKMessaging {
      *         The input parameters don't match the service's restrictions.
      * @throws ForbiddenException
      *         The client is permanently forbidden from making the request.
+     * @throws ConflictException
+     *         The request could not be processed because of conflict in the current state of the resource.
      * @throws UnauthorizedClientException
      *         The client is not currently authorized to make the request.
      * @throws ThrottledClientException
@@ -1436,6 +1438,33 @@ public interface AmazonChimeSDKMessaging {
      *      target="_top">AWS API Documentation</a>
      */
     RedactChannelMessageResult redactChannelMessage(RedactChannelMessageRequest redactChannelMessageRequest);
+
+    /**
+     * <p>
+     * Allows an <code>AppInstanceUser</code> to search the channels that they belong to. The
+     * <code>AppInstanceUser</code> can search by membership or external ID. An <code>AppInstanceAdmin</code> can search
+     * across all channels within the <code>AppInstance</code>.
+     * </p>
+     * 
+     * @param searchChannelsRequest
+     * @return Result of the SearchChannels operation returned by the service.
+     * @throws BadRequestException
+     *         The input parameters don't match the service's restrictions.
+     * @throws ForbiddenException
+     *         The client is permanently forbidden from making the request.
+     * @throws UnauthorizedClientException
+     *         The client is not currently authorized to make the request.
+     * @throws ThrottledClientException
+     *         The client exceeded its request rate limit.
+     * @throws ServiceUnavailableException
+     *         The service is currently unavailable.
+     * @throws ServiceFailureException
+     *         The service encountered an unexpected error.
+     * @sample AmazonChimeSDKMessaging.SearchChannels
+     * @see <a href="http://docs.aws.amazon.com/goto/WebAPI/chime-sdk-messaging-2021-05-15/SearchChannels"
+     *      target="_top">AWS API Documentation</a>
+     */
+    SearchChannelsResult searchChannels(SearchChannelsRequest searchChannelsRequest);
 
     /**
      * <p>