+++ aws-java-sdk-connect/src/main/java/com/amazonaws/services/connect/AbstractAmazonConnect.java
@@ -358,6 +358,11 @@ public class AbstractAmazonConnect implements AmazonConnect {
     }
 
     @Override
+    public GetCurrentUserDataResult getCurrentUserData(GetCurrentUserDataRequest request) {
+        throw new java.lang.UnsupportedOperationException();
+    }
+
+    @Override
     public GetFederationTokenResult getFederationToken(GetFederationTokenRequest request) {
         throw new java.lang.UnsupportedOperationException();
     }