+++ aws-java-sdk-connect/src/main/java/com/amazonaws/services/connect/model/TransferContactRequest.java
@@ -33,7 +33,7 @@ public class TransferContactRequest extends com.amazonaws.AmazonWebServiceReques
     private String instanceId;
     /**
      * <p>
-     * The identifier of the contact in this instance of Amazon Connect
+     * The identifier of the contact in this instance of Amazon Connect.
      * </p>
      */
     private String contactId;
@@ -104,11 +104,11 @@ public class TransferContactRequest extends com.amazonaws.AmazonWebServiceReques
 
     /**
      * <p>
-     * The identifier of the contact in this instance of Amazon Connect
+     * The identifier of the contact in this instance of Amazon Connect.
      * </p>
      * 
      * @param contactId
-     *        The identifier of the contact in this instance of Amazon Connect
+     *        The identifier of the contact in this instance of Amazon Connect.
      */
 
     public void setContactId(String contactId) {
@@ -117,10 +117,10 @@ public class TransferContactRequest extends com.amazonaws.AmazonWebServiceReques
 
     /**
      * <p>
-     * The identifier of the contact in this instance of Amazon Connect
+     * The identifier of the contact in this instance of Amazon Connect.
      * </p>
      * 
-     * @return The identifier of the contact in this instance of Amazon Connect
+     * @return The identifier of the contact in this instance of Amazon Connect.
      */
 
     public String getContactId() {
@@ -129,11 +129,11 @@ public class TransferContactRequest extends com.amazonaws.AmazonWebServiceReques
 
     /**
      * <p>
-     * The identifier of the contact in this instance of Amazon Connect
+     * The identifier of the contact in this instance of Amazon Connect.
      * </p>
      * 
      * @param contactId
-     *        The identifier of the contact in this instance of Amazon Connect
+     *        The identifier of the contact in this instance of Amazon Connect.
      * @return Returns a reference to this object so that method calls can be chained together.
      */